[ #54786 ] Rewrite the PhoneNumber matching code
This commit rewrite the most complexe and fragile part of
libQtSFLPhone. This will most probably cause regressions.

The goal old code could not be reliably be upgraded to the new
PhoneNumber and Contact merging feature. That,in turn, was
required to support more matching corner cases that are
increasingly caused by new contact related features like
history contact storing and incoming daemon features.

More code will be required to handle the various TODOs. A new
issue caused by this commit is when multiple semi-compatible URIs
are merged. The "best" one will have to be used when calling
depending of factors like the account used to call and history
data.